# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Prometheus hub, distrubed prometheus targent manager.",
        "title": "Prometheus Hub API",
        "contact": {},
        "version": "0.0.9"
    },
    "host": "localhost:8899",
    "basePath": "/",
    "paths": {
        "/ph/v1/proxy": {
            "get": {
                "description": "This endpoint proxies requests to another server.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy"
                ],
                "summary": "Reverse Proxy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Protocol (http/https)",
                        "name": "schema",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Host to proxy to",
                        "name": "host",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Port to proxy to",
                        "name": "port",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Basic auth credentials",
                        "name": "base",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bearer token for authentication",
                        "name": "bearer",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Additional parameter 1",
                        "name": "param1",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Additional parameter 2",
                        "name": "param2",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/ph/v1/selectors": {
            "get": {
                "description": "List selectors",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Selectors"
                ],
                "summary": "List selectors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "description": "Update selector by old key and value",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Selectors"
                ],
                "summary": "Update selector",
                "parameters": [
                    {
                        "description": "Update Selector Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cylonchau_pantheon_pkg_apis_query.QueryEditSelector"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_cylonchau_pantheon_pkg_apis_query.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_cylonchau_pantheon_pkg_apis_query.Response"
                        }
                    }
                }
            }
        },
...
        "/ph/v1/targets/selector/{key}/{value}": {
            "get": {
                "description": "List target with instance labels",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Targets"
                ],
                "summary": "List target with instance labels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "selector key name",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "selector value name",
                        "name": "value",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove target instance with name.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "Targets"
                ],
                "summary": "Remove prometheus target with name.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "target name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/ph/v2/targets/selector/{key}/{value}": {
            "get": {
                "description": "List target with instance labels",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Targets"
                ],
                "summary": "List target with instance labels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "selector key name",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "selector value name",
                        "name": "value",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_cylonchau_pantheon_pkg_apis_query.QueryEditSelector": {
            "type": "object",
            "required": [
                "new_key",
                "new_value",
                "old_key",
                "old_value"
            ],
            "properties": {
                "new_key": {
                    "description": "新键",
                    "type": "string"
                },
                "new_value": {
                    "description": "新值",
                    "type": "string"
                },
                "old_key": {
                    "description": "源键",
                    "type": "string"
                },
                "old_value": {
                    "description": "源值",
                    "type": "string"
                }
            }
        },
        "github_com_cylonchau_pantheon_pkg_apis_query.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}